% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encode.R
\name{decode_obj}
\alias{decode_obj}
\title{Decode hashed text into an R object}
\usage{
decode_obj(txt, compress = c("bzip2", "gzip", "xz", "none"))
}
\arguments{
\item{txt}{Hashed text.}

\item{compress}{Compression method.}
}
\description{
Decode hashed text into an R object
}
