% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{learnr_elements}
\alias{learnr_elements}
\alias{decoder_logic}
\alias{decoder_ui}
\alias{encoder_logic}
\alias{encoder_ui}
\title{Learnr addon elements}
\usage{
decoder_logic()

decoder_ui()

encoder_logic()

encoder_ui(url = "http://localhost")
}
\arguments{
\item{url}{Link url of the submission form being used.}
}
\description{
The following are addon element for learnr tutorials that enable the encoding and
decoding of hashed learnr solutions.

Note that when including these functions in a learnr Rmd document it is necessary that
the logic functions, \verb{*_logic()}, be included in an R chunk where \code{context="server"} as
they interact with the underlying Shiny functionality. Conversely, any of the ui functions,
\verb{*_ui()}, must \emph{not} be included in an R chunk with a \code{context}. Both types of functions
have been written to provide useful feedback if they detect they are in the wrong R chunk
type.
}
