% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract}
\alias{extract}
\alias{extract_hash}
\alias{extract_exercises}
\alias{extract_questions}
\title{Extract hash contents}
\usage{
extract_hash(df, hash)

extract_exercises(df, hash, include_output = FALSE)

extract_questions(df, hash, include_text = TRUE)
}
\arguments{
\item{df}{A data frame containing the hashes}

\item{hash}{The name of the column containing the hashes}

\item{include_output}{Logical. Should the exercises' code output be included.}

\item{include_text}{Logical. Should the questions' text be included.}
}
\description{
The following are helper functions for extracting data from hashed learnr solutions.
\itemize{
\item \code{extract_hash} - extracts the contents of the hashes into a list column.
\item \code{extract_questions} - extracts the contents of the hashes and expands answered questions.
\item \code{extract_exercises} - extracts the contents of the hashes and expands answered exercises.
}
}
