% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract}
\alias{extract}
\alias{extract_hash}
\alias{extract_exercises}
\alias{extract_questions}
\title{Extract hash contents}
\usage{
extract_hash(df, hash = "hash")

extract_exercises(df, hash = "hash")

extract_questions(df, hash = "hash")
}
\arguments{
\item{df}{Data Frame. A data frame containing hash in a character column.}

\item{hash}{Character. The name of the column containing the hashes}
}
\description{
The following are helper functions for extracting data from hashed learnr solutions.
\itemize{
\item \code{extract_hash} - extracts the contents of the hashes into label, type, answer, correct, and timestamp columns
\item \code{extract_questions} - extracts the contents of the hashes for answered questions.
\item \code{extract_exercises} - extracts the contents of the hashes for answered exercises.
}
}
